+++ app/src/main/kotlin/io/celox/flipperripper/di/Qualifiers.kt
@@ -21,3 +21,7 @@ annotation class OnDeviceEngine
 @Qualifier
 @Retention(AnnotationRetention.BINARY)
 annotation class RemoteEngine
+
+@Qualifier
+@Retention(AnnotationRetention.BINARY)
+annotation class WebViewEngine